Split store into data and authentication
